requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cssselect==1.1.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.0.7
django-heroku==0.3.1
djangorestframework==3.11.0
factory-boy==2.12.0
fake-useragent==0.1.11
Faker==4.1.0
filelock==3.0.12
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.1
lxml==4.5.1
mock==4.0.2
mysqlclient==1.4.6
parsel==1.6.0
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
selectorlib==0.16.0
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
tzlocal==2.1
urllib3==1.25.9
useragent==0.1.1
virtualenv==20.0.21
w3lib==1.22.0
whitenoise==5.1.0
zipp==3.1.0
